+++ bundles/org.palladiosimulator.simulizar.failurescenario/src/org/palladiosimulator/simulizar/failurescenario/di/FailurescenarioExtensionComponent.java
@@ -6,8 +6,8 @@ import org.palladiosimulator.simulizar.di.component.core.SimuLizarRuntimeCompone
 import org.palladiosimulator.simulizar.di.component.dependency.SimEngineComponent;
 import org.palladiosimulator.simulizar.di.component.dependency.SimuComFrameworkComponent;
 import org.palladiosimulator.simulizar.di.extension.ExtensionComponent;
+import org.palladiosimulator.simulizar.di.scopes.RuntimeExtensionScope;
 import org.palladiosimulator.simulizar.failurescenario.interpreter.FailurescenarioObserver;
-import org.palladiosimulator.simulizar.scopes.RuntimeExtensionScope;
 
 import dagger.Component;
