+++ bundles/org.palladiosimulator.simulizar.failurescenario/src/org/palladiosimulator/simulizar/failurescenario/di/FailurescenarioExtensionModule.java
@@ -1,7 +1,7 @@
 package org.palladiosimulator.simulizar.failurescenario.di;
 
+import org.palladiosimulator.simulizar.di.scopes.RuntimeExtensionScope;
 import org.palladiosimulator.simulizar.interpreter.InterpreterDefaultContext;
-import org.palladiosimulator.simulizar.scopes.RuntimeExtensionScope;
 
 import dagger.Binds;
 import dagger.Module;
