+++ bundles/org.palladiosimulator.simulizar.failurescenario/src/org/palladiosimulator/simulizar/failurescenario/di/FailurescenarioRootComponent.java
@@ -5,10 +5,10 @@ import org.eclipse.core.runtime.IExecutableExtensionFactory;
 import org.palladiosimulator.simulizar.core.runconfig.SimuLizarWorkflowConfiguration;
 import org.palladiosimulator.simulizar.di.component.core.SimuLizarRootComponent;
 import org.palladiosimulator.simulizar.di.extension.ExtensionComponent;
+import org.palladiosimulator.simulizar.di.scopes.RootExtensionScope;
 import org.palladiosimulator.simulizar.failurescenario.jobs.FailurescenarioModelContribution;
 import org.palladiosimulator.simulizar.failurescenario.jobs.config.LoadFailurescenarioExtensionIntoBlackboardJobConfig;
 import org.palladiosimulator.simulizar.failurescenario.ui.config.FailurescenarioExtensionFileInputConfigurationBuilder;
-import org.palladiosimulator.simulizar.scopes.RootExtensionScope;
 
 import dagger.Component;
 import dagger.Provides;
